New debug functionality, new Tech Tree virtual graph
The debug functionality allows you to print out files as a sort of test
during runtime. One function in debug allows you to quickly add the
localtime to the debug log's filename, however, this function hasn't
been tested.

The Tech Tree virtual graph allows you to examine the requirements and
costs of unlocking a given unit.